+++ CHANGELOG.md
@@ -8,6 +8,7 @@ develop
 ### Fixes
 
 - Fix an issue where application could rarely freeze after replace large document.
+- Fix an issue where new syntax style could't be created.
 - Fix an issue where new value of last edited text field in preferences was occasionally discarded.
 - Fix error message of syntax style valitation.